added first api data

+++ lib/models/currentWeatherModel.dart
@@ -1,3 +1,4 @@
+import 'dart:core';
 // To parse this JSON data, do
 //
 //     final currentWeatherData = currentWeatherDataFromJson(jsonString);
@@ -53,9 +54,9 @@ class Clouds {
 }
 
 class Main {
-  double? temp;
-  double? tempMin;
-  double? tempMax;
+  int? temp;
+  int? tempMin;
+  int? tempMax;
   int? humidity;
   int? seaLevel;
   int? grndLevel;
@@ -70,9 +71,9 @@ class Main {
   });
 
   factory Main.fromJson(Map<String, dynamic> json) => Main(
-        temp: json["temp"]?.toDouble(),
-        tempMin: json["temp_min"]?.toDouble(),
-        tempMax: json["temp_max"]?.toDouble(),
+        temp: json["temp"]?.toInt(),
+        tempMin: json["temp_min"]?.toInt(),
+        tempMax: json["temp_max"]?.toInt(),
         humidity: json["humidity"],
         seaLevel: json["sea_level"],
         grndLevel: json["grnd_level"],
